Very Simple, Child Check In Mock Up, With SQLite
This is just a very simple mock up program of a child check in interface
including SQLite. I wrote this mainly as a learning tool to better
understand C#. Figured I should post it to development, to show some
evidence of my learning in case anyone cares. For simplicity, the only
current ID's in the database are 1, 2 , and 3.