Login window and admin window changes
can login now with user name "999999" and password "9999".  We will need
to redesign later to store user names and passwords as strings in the
database for managers only.  Parents will still use their IDs and PIN
from the parent side of the application.

Also made it possible to logout of the system and go back to the
"MainWindow"