Further fleshed out child check in window with displays
Added a ListBox to show selection of logged in children. Added parent
name and picture display. Added a date and time display. Added a drop
down menu for events, and insured that an event choice is chosen before
login is allowed. Can be tested using ID: 111111 and PIN: 1111 or ID:
222222 and PIN: 2222 or ID: 333333 and PIN: 3333. All displays should be
working, but behind the scenes actual login functionality/record storing
is not functional yet.