database schema + sqlite database
Put up the database schema and put a sqlite database in the Database
folder that follows the schema almost exactly.  I renamed the
transactions table to be the event data table because its a more intent
revealing name.  I also changed the start and end date to be a date and
day field.  Both are nullable.  We can change this later if we want, but
I realized that a start/end date will not really work since there are
weekly events.  Everyone should use this database until we can get a
MySQL database running.