Further child check in functionality
Children can now be checked in. The transaction will be recorded in the
database including date , check in time, event id, and connection id.
Children can also be checked back out and the check out time will be
recorded in the appropriate database record. The check in and check out
listboxes should populate correctly on log in, accounting for both
children already logged in, and those that are not. Total transaction
cost is not yet calculated/recorded.