Made Forms function (mostly)
Parent and Child forms now correctly populate with information from the
database. You can use the submit buttons to change their values in the
database. You can use the delete button to delete  parents and childern
from the database.  There is a next button that you can click on
editChilderen that cycles through the childeren that the parent can see.