Parent Report and Business Report
The Parent Report and Business Report windows are now basically done as
far as functionality is concerned.  There is, however, almost no error
checking and one or two small things still need to be implemented.  The
database is going to need a lot more information in it to fully test all
these things, but I will speak with Blaine about it on Monday.  Also, we
have not yet decided how to handle current due balance.  This is
something we will have to make a modification to the database for,
however I do not believe the necessary modification should break any
code.

Brian: I added a ParentInfoDB class.  You may be able to use it for some
of the stuff you do, but I have not fully looked through your code so I
am not sure if it will be useful to you or not.  In any case, there are
plenty of MySQL DB connection and reading examples in it.