Convert to MySQL
So I am having big probmles connecting to the vpn thing. I cant seam to
egt it to work so I looked at your guyses code for connecting to the
server and I changed my code to do MySQL not SQLite. I was not able to
test it however because I cant connect to the server. The old code is
commented out so I can change it back if we have to for presentation.