Settings window and functionality
I added a settings window for the parent side. This window allows an
admin to change the database connection details, and change a few
general details such as max monthly rate, photo locations, and billing
period. This information is saved into a config file(settings.txt). This
information must remain consistent and be accessed from several points,
so I implemented a singleton pattern to contain it during runtime. The
singleton will be loaded once at startup with details from the file. I
also used the file picker idea that was suggested by Brain. The settings
window contains a folder picker to choose the default picture path(if
you guys test it, just type in our relative path manually or else full
paths will be stored in the database). The edit child/parent widows
contain file pickers that start from that default path, so the user can
easily choose an image. Also, if no config file currently exists, the
user will be prompted to configure settings before use, and a file will
be created.