Created the initial project window again
I also needed to change the name of "ParentLogin" on the manager side of
things to be "ParentIDEntry" since that name fits better and there was a
"ParentLogin" window on the parent use side of things where the parent
actually logs into the program to check in/out children.

The program also now compiles and runs.