Diagram Updates
Made new sequence and state diagrams for my code, that reflect the new
project structure. If you guys want we can do a combined state diagram
on Tuesday for the whole project; but I think that will get pretty messy
and it will probably be allot easier to talk about what we did
individually, if we keep those seperate.