admin login and parent report changes
I made the admin login look a little bit nicer and fixed the tab
indexing so its easier to login without selecting the boxes.  I also had
to do a fair amount to work around the fact that the version of SQLite
we are using does not seem to have a function to format float values to
two decimal places.