SQLite output formatting work around #2
I had to do the formatting for the total cost per parent differently in
the Business report due to the sum and group by functions that are used
in that sql statement.  After struggling to get the formatting done in
the querry, I finally switched to trying to do it on the C# side of
things and got it to work that way.