More Business Report Functionality and ease of use
Setup the Business Report window to actually load the current month
report from the current month and fixed some bugs in the Business Report
window.

Added several things to make it easier to navigate the program,
primarily by allowing hitting enter while in a text field to either move
you to the next needed entry or to submit your entry.  Also set the
windows up so if you go to the admin side of the program, the user
selection closes and the same for the parent side.