Reports bug fix and event adding
It is not possible to edit an event right now as it will simply add a
new event with the data you enter, and there are a few display bugs
involved with hitting the edit event button, but no runtime errors as
far as I can tell.

When we next meet up ill want to discuss another database change for the
events table, which will affect how I would solve the current bugs.