bug fixes, code conversion, and dll change
I fixed some bugs related to switching to the newer version of the
database, converted my code to use the printf function and fixed another
bug in the process, and updated the dll for sqlite in the database
folder to be for our current database version.

I was closing the database connection in a destructor in one place, but
with the new version there was a crash.  I changed the database opening
and closing to be more standard to fix the problem.

The old dll is still in the folder with an added .cur on the end just in
case we need it at some point for some reason.