Bug fixes, user friendly, and admin access level
Fixed several windows to be more user friendly, fixed the bug with
negative payment values, and added a proof on concept for access level.

Currently the add/edit events button is not available for level 2
admins.  What we leave available for level 2 admins will ultimately be
up to our client, but I wanted a proof of concept for our demonstration.