dummy data and a business reports bug fix
Modified my script writer and used the sqlite3.exe program to insert
around 16k records into the database.  With all of this dummy data, I
found a bug in the business reports and fixed it.

Currently the dummy data is all for a specific day in the February
billing cycle, however at some point i will modify the script writer to
create entrees for random dates anytime in the last year or so.

I copied the database before inserting all the dummy data and renamed it
to include "nodummy" in the name, then named the database will all the
dummy data to be the database we are currently reading from.  You can
switch out the databases if you need to in this manner, but please do
not delete any databases (it took around 15 minutes for the command
prompt to finish entering all the sql statements I copy and pasted into
it).