Hashing fun
Added the hashing check for admin login
Added a step to hash the parents pin when creating a new parent
Added code o the parent login side to hash the PIN before loging in(This
was commented out Brandon so you can still login as a parent)
We will need to either wipe the database and add only parents whose PINs
we hash or put in the correct hash for all of the PINs currently in the
database

I did not add hashing for a new admin password. I wasnt sure if Blaine
had finished with his work.