Num lock bug, and Spokane Club Icon
Fixed a bug in the guardian login when num lock is pressed. Also added a
Spokane Club picture as our project icon for now.

Note: Icons do not show in debug mode, so you will not see it when
running the project in Visual Studio. You can see it when just running
it from the exe(like I do for our presentations).