Last of PDF stuff
PDF functionality is now fully operational.  User can select the name to
save the file as and .pdf will be added if the user does not add it.
The savefile dialog will also open in the directory where the user has
selected that reports should be saved to in the settings window.  If
there are fields which are too long in the report, they are cut off to
avoid one field writing over the top of another field.