Connected help button to user manual
The help now has the functionality to open our user manual(currently
just the blank stand in).

Note: It is set up to open the file in
"C:/Users/username/Documents/Childcare Application", so a copy of the
PDF must be placed there for it to work(This will be done automatically
by the installer).