Fixed bug created by earlier changes.
A database close statement was removed at some point in the changes
earlier today, and caused the database to stay locked up when a child
was unlinked. I re-added the close statement. Didn't see any more cases
of this, but it looks like allot of similar changes were made(careful
guys).